Initial commit
The Debian repository is already checked, to some extent. Currently, it
downloads the Packages files, parses them, extracts the SHA1 checksum and the
file URL, downloads the file, produces progress information about that
download, and will produce an "ALERT" if a file checksum does not match,
or an "OK" if it does.

Still to do:
- GPG verification of the Release file
- Checking whether the Release file's checksum of the Packages file
  matches
- The whole RPM side of things (XML, yuck).